Initial snapcraft yaml
The following will build and compile 4 different compilers for the 2
different architectures. Each 4 compilers target a different
architecture that we can cross compile to. We build the 4 cross
compilers for amd64 and arm64, because these are the mose common
development environment architectures. If we require others, then we can
add them at a later date.

We currently don't have riscv working atm, but that can be added at a
later stage.